+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
-## 0.1 (2019-06-??)
+## 0.2 (2019-06-24)
+- merge tag and dist commands into tagdist
+- if sdist fails then reset tag
+- drop push-tags command
+
+
+## 0.1 (2019-06-24)
 - add published lock file: prevents tagging/dist/upload after publish
 - tag before sdist if no tag
 - push tag after publishing